+++ src/twg2/treeLike/parameters/IndexedTreeTraverseParameters.java
@@ -1,11 +1,11 @@
 package twg2.treeLike.parameters;
 
-import java.util.List;
 import java.util.function.Function;
 import java.util.function.IntConsumer;
 import java.util.function.Predicate;
 
 import lombok.val;
+import twg2.collections.interfaces.ListReadOnly;
 import twg2.treeLike.IndexedSubtreeConsumer;
 import twg2.treeLike.TreeTraversalOrder;
 
@@ -14,7 +14,7 @@ import twg2.treeLike.TreeTraversalOrder;
  * @since 2015-8-31
  * @param <T> the type of data in each tree node
  */
-public class IndexedTreeTraverseParameters<T> extends AbstractTreeTraverseParameters<T, T, List<? extends T>, IndexedSubtreeConsumer<T>> {
+public class IndexedTreeTraverseParameters<T> extends AbstractTreeTraverseParameters<T, T, ListReadOnly<? extends T>, IndexedSubtreeConsumer<T>> {
 	// package-private
 	IndexedSubtreeConsumer<T> consumer;
 
@@ -29,20 +29,20 @@ public class IndexedTreeTraverseParameters<T> extends AbstractTreeTraverseParame
 	@Override
 	public IndexedTreeTraverseParameters<T> setHasChildren(Predicate<T> hasChildren) { super.setHasChildren(hasChildren); return this; }
 	@Override
-	public IndexedTreeTraverseParameters<T> setChildrenGetter(Function<T, List<? extends T>> childrenGetter) { super.setChildrenGetter(childrenGetter); return this; }
+	public IndexedTreeTraverseParameters<T> setChildrenGetter(Function<T, ListReadOnly<? extends T>> childrenGetter) { super.setChildrenGetter(childrenGetter); return this; }
 	@Override
 	public IndexedTreeTraverseParameters<T> setStartSubtreeFunc(IntConsumer startSubtreeFunc) { super.setStartSubtreeFunc(startSubtreeFunc); return this; }
 	@Override
 	public IndexedTreeTraverseParameters<T> setEndSubtreeFunc(IntConsumer endSubtreeFunc) { super.setEndSubtreeFunc(endSubtreeFunc); return this; }
 
 
-	public IndexedTreeTraverseParameters(T tree, boolean onlyVisitLeaves, TreeTraversalOrder traversalOrder, Predicate<T> hasChildren, Function<T, List<? extends T>> childrenGetter,
+	public IndexedTreeTraverseParameters(T tree, boolean onlyVisitLeaves, TreeTraversalOrder traversalOrder, Predicate<T> hasChildren, Function<T, ListReadOnly<? extends T>> childrenGetter,
 			IndexedSubtreeConsumer<T> consumer, IntConsumer startSubtreeFunc, IntConsumer endSubtreeFunc) {
 		this(tree, onlyVisitLeaves, false, traversalOrder, hasChildren, childrenGetter, consumer, startSubtreeFunc, endSubtreeFunc);
 	}
 
 
-	public IndexedTreeTraverseParameters(T tree, boolean onlyVisitLeaves, boolean skipNullRoot, TreeTraversalOrder traversalOrder, Predicate<T> hasChildren, Function<T, List<? extends T>> childrenGetter,
+	public IndexedTreeTraverseParameters(T tree, boolean onlyVisitLeaves, boolean skipNullRoot, TreeTraversalOrder traversalOrder, Predicate<T> hasChildren, Function<T, ListReadOnly<? extends T>> childrenGetter,
 			IndexedSubtreeConsumer<T> consumer, IntConsumer startSubtreeFunc, IntConsumer endSubtreeFunc) {
 		super(tree, onlyVisitLeaves, skipNullRoot, traversalOrder, hasChildren, childrenGetter, startSubtreeFunc, endSubtreeFunc);
 		this.consumer = consumer;
@@ -72,19 +72,19 @@ public class IndexedTreeTraverseParameters<T> extends AbstractTreeTraverseParame
 	}
 
 
-	public static <_T> IndexedTreeTraverseParameters<_T> of(_T tree, boolean onlyVisitLeaves, TreeTraversalOrder traversalOrder, Predicate<_T> hasChildren, Function<_T, List<? extends _T>> childrenGetter) {
+	public static <_T> IndexedTreeTraverseParameters<_T> of(_T tree, boolean onlyVisitLeaves, TreeTraversalOrder traversalOrder, Predicate<_T> hasChildren, Function<_T, ListReadOnly<? extends _T>> childrenGetter) {
 		val params = new IndexedTreeTraverseParameters<_T>(tree, onlyVisitLeaves, traversalOrder, hasChildren, childrenGetter, null, null, null);
 		return params;
 	}
 
 
-	public static <_T> IndexedTreeTraverseParameters<_T> leafNodes(_T tree, TreeTraversalOrder traversalOrder, Predicate<_T> hasChildren, Function<_T, List<? extends _T>> childrenGetter) {
+	public static <_T> IndexedTreeTraverseParameters<_T> leafNodes(_T tree, TreeTraversalOrder traversalOrder, Predicate<_T> hasChildren, Function<_T, ListReadOnly<? extends _T>> childrenGetter) {
 		val params = new IndexedTreeTraverseParameters<_T>(tree, true, traversalOrder, hasChildren, childrenGetter, null, null, null);
 		return params;
 	}
 
 
-	public static <_T> IndexedTreeTraverseParameters<_T> allNodes(_T tree, TreeTraversalOrder traversalOrder, Predicate<_T> hasChildren, Function<_T, List<? extends _T>> childrenGetter) {
+	public static <_T> IndexedTreeTraverseParameters<_T> allNodes(_T tree, TreeTraversalOrder traversalOrder, Predicate<_T> hasChildren, Function<_T, ListReadOnly<? extends _T>> childrenGetter) {
 		val params = new IndexedTreeTraverseParameters<_T>(tree, false, traversalOrder, hasChildren, childrenGetter, null, null, null);
 		return params;
 	}