+++ src/twg2/treeLike/parameters/TreePathTraverseParameters.java
@@ -1,11 +1,10 @@
 package twg2.treeLike.parameters;
 
-import java.util.Collection;
 import java.util.function.Function;
 import java.util.function.IntConsumer;
 import java.util.function.Predicate;
 
-import lombok.val;
+import twg2.collections.interfaces.ListReadOnly;
 import twg2.treeLike.TreePathConsumer;
 import twg2.treeLike.TreeTraversalOrder;
 
@@ -14,7 +13,7 @@ import twg2.treeLike.TreeTraversalOrder;
  * @since 2015-9-1
  * @see AbstractTreeTraverseParameters
  */
-public class TreePathTraverseParameters<T> extends AbstractTreeTraverseParameters<T, T, Collection<T>, TreePathConsumer<T>> {
+public class TreePathTraverseParameters<T> extends AbstractTreeTraverseParameters<T, T, ListReadOnly<T>, TreePathConsumer<T>> {
 	// package-private
 	TreePathConsumer<T> consumer;
 
@@ -29,7 +28,7 @@ public class TreePathTraverseParameters<T> extends AbstractTreeTraverseParameter
 	@Override
 	public TreePathTraverseParameters<T> setHasChildren(Predicate<T> hasChildren) { super.setHasChildren(hasChildren); return this; }
 	@Override
-	public TreePathTraverseParameters<T> setChildrenGetter(Function<T, Collection<T>> childrenGetter) { super.setChildrenGetter(childrenGetter); return this; }
+	public TreePathTraverseParameters<T> setChildrenGetter(Function<T, ListReadOnly<T>> childrenGetter) { super.setChildrenGetter(childrenGetter); return this; }
 	@Override
 	public TreePathTraverseParameters<T> setStartSubtreeFunc(IntConsumer startSubtreeFunc) { super.setStartSubtreeFunc(startSubtreeFunc); return this; }
 	@Override
@@ -37,9 +36,9 @@ public class TreePathTraverseParameters<T> extends AbstractTreeTraverseParameter
 
 
 	@SuppressWarnings("unchecked")
-	public TreePathTraverseParameters(T tree, boolean onlyVisitLeaves, TreeTraversalOrder traversalOrder, Predicate<T> hasChildren, Function<T, ? extends Collection<T>> childrenGetter,
+	public TreePathTraverseParameters(T tree, boolean onlyVisitLeaves, TreeTraversalOrder traversalOrder, Predicate<T> hasChildren, Function<T, ? extends ListReadOnly<T>> childrenGetter,
 			TreePathConsumer<T> consumer, IntConsumer startSubtreeFunc, IntConsumer endSubtreeFunc) {
-		super(tree, onlyVisitLeaves, false, traversalOrder, hasChildren, (Function<T, Collection<T>>) childrenGetter, startSubtreeFunc, endSubtreeFunc);
+		super(tree, onlyVisitLeaves, false, traversalOrder, hasChildren, (Function<T, ListReadOnly<T>>) childrenGetter, startSubtreeFunc, endSubtreeFunc);
 		this.consumer = consumer;
 	}
 
@@ -67,21 +66,18 @@ public class TreePathTraverseParameters<T> extends AbstractTreeTraverseParameter
 	}
 
 
-	public static <_T> TreePathTraverseParameters<_T> of(_T tree, boolean onlyVisitLeaves, TreeTraversalOrder traversalOrder, Predicate<_T> hasChildren, Function<_T, ? extends Collection<_T>> childrenGetter) {
-		val params = new TreePathTraverseParameters<_T>(tree, onlyVisitLeaves, traversalOrder, hasChildren, childrenGetter, null, null, null);
-		return params;
+	public static <_T> TreePathTraverseParameters<_T> of(_T tree, boolean onlyVisitLeaves, TreeTraversalOrder traversalOrder, Predicate<_T> hasChildren, Function<_T, ? extends ListReadOnly<_T>> childrenGetter) {
+		return new TreePathTraverseParameters<_T>(tree, onlyVisitLeaves, traversalOrder, hasChildren, childrenGetter, null, null, null);
 	}
 
 
-	public static <_T> TreePathTraverseParameters<_T> leafNodes(_T tree, TreeTraversalOrder traversalOrder, Predicate<_T> hasChildren, Function<_T, ? extends Collection<_T>> childrenGetter) {
-		val params = new TreePathTraverseParameters<_T>(tree, true, traversalOrder, hasChildren, childrenGetter, null, null, null);
-		return params;
+	public static <_T> TreePathTraverseParameters<_T> leafNodes(_T tree, TreeTraversalOrder traversalOrder, Predicate<_T> hasChildren, Function<_T, ? extends ListReadOnly<_T>> childrenGetter) {
+		return new TreePathTraverseParameters<_T>(tree, true, traversalOrder, hasChildren, childrenGetter, null, null, null);
 	}
 
 
-	public static <_T> TreePathTraverseParameters<_T> allNodes(_T tree, TreeTraversalOrder traversalOrder, Predicate<_T> hasChildren, Function<_T, ? extends Collection<_T>> childrenGetter) {
-		val params = new TreePathTraverseParameters<_T>(tree, false, traversalOrder, hasChildren, childrenGetter, null, null, null);
-		return params;
+	public static <_T> TreePathTraverseParameters<_T> allNodes(_T tree, TreeTraversalOrder traversalOrder, Predicate<_T> hasChildren, Function<_T, ? extends ListReadOnly<_T>> childrenGetter) {
+		return new TreePathTraverseParameters<_T>(tree, false, traversalOrder, hasChildren, childrenGetter, null, null, null);
 	}
 
 }